requirements.txt
# requirements.txt

prefect==3.0.3
pandas==2.2.3
requests==2.32.3
python-telegram-bot==21.6
python-dotenv==1.0.1
psutil==6.0.0
matplotlib==3.9.2
